转付确认接口修改

+++ src/main/java/com/hand/hls/partner/dto/TermRepayDetailApplyDTO.java
@@ -13,15 +13,15 @@ public class TermRepayDetailApplyDTO extends BaseDTO {
 
 
     @NotNull(message = "期次号不能为空")
-    private Integer termNo;//期次号
+    private Long termNo;//期次号
     @NotNull(message = "实际还款本金不能为空")
-    private Long repayPrincipal;//实际还款本金；单位 分
+    private Double repayPrincipal;//实际还款本金；单位 分
     @NotNull(message = "实际还款利息不能为空")
-    private Long repayInterest;//实际还款利息；单位 分
+    private Double repayInterest;//实际还款利息；单位 分
     @NotNull(message = "实际还款罚息不能为空")
-    private Long repayPenalty;//实际还款罚息；单位 分
+    private Double repayPenalty;//实际还款罚息；单位 分
     @NotNull(message = "实际还款总金额不能为空")
-    private Long repayAmount;//实际还款总金额；单位 分
+    private Double repayAmount;//实际还款总金额；单位 分
     private String transactionNo;//结算单号
     private String externalDeductNo;//代扣交易单号
